requirements.txt
flask
gevent
python-dotenv
psycopg2
flask_cors
oracledb
hdbcli
